+++ lib/models/social_user_model.dart
